+++ core/src/main/java/org/seedstack/seed/core/internal/init/ConsoleManager.java
@@ -5,19 +5,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.core.internal.init;
 
-import java.io.FilterOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.io.PrintStream;
-import java.io.UnsupportedEncodingException;
-import java.nio.charset.Charset;
 import org.fusesource.jansi.AnsiConsole;
-import org.fusesource.jansi.AnsiOutputStream;
+import org.fusesource.jansi.AnsiPrintStream;
+import org.fusesource.jansi.FilterPrintStream;
+import org.fusesource.jansi.WindowsAnsiPrintStream;
+import org.fusesource.jansi.internal.CLibrary;
 import org.seedstack.seed.ApplicationConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ConsoleManager {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ConsoleManager.class);
     private final PrintStream savedOut = System.out;
     private final PrintStream savedErr = System.err;
 
@@ -30,8 +32,8 @@ public class ConsoleManager {
     }
 
     public synchronized void install(ApplicationConfig.ColorOutput colorOutput) {
-        System.setOut(wrapPrintStream(System.out, colorOutput));
-        System.setErr(wrapPrintStream(System.err, colorOutput));
+        System.setOut(wrapPrintStream(System.out, FileNo.STDOUT, colorOutput, true));
+        System.setErr(wrapPrintStream(System.err, FileNo.STDERR, colorOutput, false));
     }
 
     public synchronized void uninstall() {
@@ -39,49 +41,84 @@ public class ConsoleManager {
         System.setErr(savedErr);
     }
 
-    private PrintStream wrapPrintStream(final PrintStream printStream, ApplicationConfig.ColorOutput colorOutput) {
-        OutputStream outputStream;
+    private PrintStream wrapPrintStream(final PrintStream inputStream, FileNo fileno,
+            ApplicationConfig.ColorOutput colorOutput, boolean log) {
         try {
-            if (colorOutput == ApplicationConfig.ColorOutput.PASSTHROUGH || Boolean.getBoolean("jansi.passthrough")) {
-                outputStream = printStream;
-            } else if (colorOutput == ApplicationConfig.ColorOutput.DISABLE || Boolean.getBoolean("jansi.strip")) {
-                outputStream = basicOutput(printStream);
+            if (colorOutput == ApplicationConfig.ColorOutput.PASSTHROUGH) {
+                if (log) {
+                    LOGGER.info("Color output passthrough: leaving stdout and stderr untouched");
+                }
+                return inputStream;
+            } else if (colorOutput == ApplicationConfig.ColorOutput.DISABLE) {
+                if (log) {
+                    LOGGER.info("Color output disabled: stripping stdout and stderr of ANSI codes");
+                }
+                return strippedOutput(inputStream);
             } else if (colorOutput == ApplicationConfig.ColorOutput.ENABLE) {
-                outputStream = ansiOutput(printStream);
+                if (log) {
+                    LOGGER.info("Color output enabled: allowing ANSI codes on stdout and stderr");
+                }
+                return ansiOutput(inputStream);
             } else if (colorOutput == ApplicationConfig.ColorOutput.AUTODETECT) {
                 if (isXtermColor()) {
                     // enable color in recognized XTERM color modes
-                    outputStream = ansiOutput(printStream);
+                    if (log) {
+                        LOGGER.info("XTERM detected: allowing ANSI codes on stdout and stderr");
+                    }
+                    return ansiOutput(inputStream);
                 } else if (isIntelliJ()) {
                     // enable color under Intellij
-                    outputStream = ansiOutput(printStream);
+                    if (log) {
+                        LOGGER.info("IntelliJ detected: allowing ANSI codes on stdout and stderr");
+                    }
+                    return ansiOutput(inputStream);
                 } else {
                     // let Jansi handle other detection
-                    outputStream = AnsiConsole.wrapOutputStream(printStream);
+                    PrintStream result;
+                    switch (fileno) {
+                        case STDOUT:
+                            result = AnsiConsole.wrapPrintStream(inputStream, CLibrary.STDOUT_FILENO);
+                            break;
+                        case STDERR:
+                            result = AnsiConsole.wrapPrintStream(inputStream, CLibrary.STDERR_FILENO);
+                            break;
+                        default:
+                            result = strippedOutput(inputStream);
+                            break;
+                    }
+                    if (log) {
+                        if (result instanceof WindowsAnsiPrintStream) {
+                            LOGGER.info("Windows console detected: using native coloring for stdout and stderr");
+                        } else if (result instanceof AnsiPrintStream) {
+                            LOGGER.info("ANSI not supported: stripping stdout and stderr of ANSI codes");
+                        } else {
+                            LOGGER.info("ANSI supported: allowing ANSI codes on stdout and stderr");
+                        }
+                    }
+                    return result;
                 }
             } else {
                 // Fallback to stripping ANSI codes
-                outputStream = basicOutput(printStream);
+                if (log) {
+                    LOGGER.info("Fallback to stripping ANSI codes");
+                }
+                return strippedOutput(inputStream);
             }
-        } catch (Throwable e) {
+        } catch (Throwable e1) {
             // If any error occurs, strip ANSI codes
-            outputStream = basicOutput(printStream);
-        }
-
-        try {
-            return new PrintStream(outputStream, false, Charset.defaultCharset().name());
-        } catch (UnsupportedEncodingException e) {
-            // if for some reason the default encoding is unsupported, fallback to passthrough
-            return printStream;
+            if (log) {
+                LOGGER.info("Error loading Jansi (unsupported platform ?): stripping stdout and stderr of ANSI codes");
+            }
+            return strippedOutput(inputStream);
         }
     }
 
-    private FilterOutputStream basicOutput(PrintStream stream) {
-        return new AnsiOutputStream(stream);
+    private AnsiPrintStream strippedOutput(PrintStream stream) {
+        return new AnsiPrintStream(stream);
     }
 
-    private FilterOutputStream ansiOutput(PrintStream stream) {
-        return new ColorOutputStream(stream);
+    private ColorPrintStream ansiOutput(PrintStream stream) {
+        return new ColorPrintStream(stream);
     }
 
     private boolean isXtermColor() {
@@ -108,16 +145,21 @@ public class ConsoleManager {
         private static final ConsoleManager INSTANCE = new ConsoleManager();
     }
 
-    private static class ColorOutputStream extends FilterOutputStream {
-        private ColorOutputStream(OutputStream stream) {
+    private static class ColorPrintStream extends FilterPrintStream {
+        private ColorPrintStream(PrintStream stream) {
             super(stream);
         }
 
         @Override
-        public void close() throws IOException {
-            write(AnsiOutputStream.RESET_CODE);
-            flush();
+        public void close() {
+            ps.print(AnsiPrintStream.RESET_CODE);
+            ps.flush();
             super.close();
         }
+    }
+
+    private enum FileNo {
+        STDOUT,
+        STDERR
     }
 }