+++ core/src/test/java/org/seedstack/seed/core/CryptDemo.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.seed.core;
+
+public class CryptDemo {
+    public static void main(String[] args) throws Exception {
+        Seed.getToolLauncher("crypt").launch(new String[]{"changeMe"});
+    }
+}