+++ specs/src/main/java/org/seedstack/seed/Bind.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed;
 
 import java.lang.annotation.Documented;
@@ -20,10 +21,10 @@ import java.lang.annotation.Target;
  *
  * <pre>
  * {@literal @}Bind
- * public class SomeImplementation {...}
+ *  public class SomeImplementation {...}
  *
  * {@literal @}Inject
- * SomeImplementation someImplementation;
+ *  SomeImplementation someImplementation;
  * </pre>
  *
  * <p>
@@ -33,30 +34,41 @@ import java.lang.annotation.Target;
  *
  * <pre>
  * {@literal @}Bind(from = SomeInterface.class)
- * public class SomeImplementation implements SomeInterface {...}
+ *  public class SomeImplementation implements SomeInterface {...}
  *
  * {@literal @}Inject
- * SomeInterface someInterface;
+ *  SomeInterface someInterface;
  * </pre>
  *
  * <p>
- * When an injection class is specified and a qualifier annotation is present on the implementation class, it is used to
- * further refine the injection key:
+ * <strong>The {@link Bind} annotation allows to override any interface-based SeedStack binding.</strong> To do this,
+ * create a custom implementation of the SeedStack interface you want to customize (this is the type you use at the
+ * injection point) and set the {@link Bind#override()} boolean to true. Your custom implementation will replace the
+ * SeedStack one.
+ * </p>
+ * <p></p>
+ * <p>
+ * When a qualifier annotation is present on the implementation class, it is used to make the injection point more
+ * specific:
  * </p>
  *
  * <pre>
  * {@literal @}Qualifier
  * {@literal @}Retention(RetentionPolicy.RUNTIME)
- * public interface {@literal @}SomeQualifier {...}
+ *  public interface {@literal @}SomeQualifier {...}
  *
  * {@literal @}Bind(from = SomeInterface.class)
  * {@literal @}SomeQualifier
- * public class SomeImplementation implements SomeInterface {...}
+ *  public class SomeImplementation implements SomeInterface {...}
  *
  * {@literal @}Inject
  * {@literal @}SomeQualifier
- * SomeInterface someInterface;
+ *  SomeInterface someInterface;
  * </pre>
+ *
+ * <p>When having multiple implementations of the same interface, using a different qualifier on each implementation
+ * allows to create multiple bindings. You can then choose the implementation by specifying the corresponding qualifier
+ * at injection point.</p>
  */
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)