+++ web/undertow/src/test/java/org/seedstack/seed/undertow/RefreshIT.java
@@ -5,6 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.seed.undertow;
 
 import com.google.inject.Injector;
@@ -59,7 +60,7 @@ public class RefreshIT {
 
     private ResponseSpecification expect() {
         return RestAssured.given()
-                .config(RestAssured.config().sslConfig(SSLConfig.sslConfig().relaxedHTTPSValidation("SSL")))
+                .config(RestAssured.config().sslConfig(SSLConfig.sslConfig().relaxedHTTPSValidation()))
                 .expect();
     }
 }